refactor: update and clean up code
- create utils file and move common func there
- add validation on extension.yaml
- payload size check before adding to bulk.
