+++ functions/src/config.ts
@@ -1,3 +1,4 @@
+'use strict';
 /*
  * Copyright 2021 Algolia
  *
@@ -15,9 +16,9 @@
  */
 
 export default {
-  algoliaAppId: process.env.ALGOLIA_APP_ID as string,
-  algoliaAPIKey: process.env.ALGOLIA_API_KEY as string,
-  algoliaIndexName: process.env.ALGOLIA_INDEX_NAME as string,
-  collectionPath: process.env.COLLECTION_PATH as string,
-  fields: process.env.FIELDS as string,
+  algoliaAppId: process.env.ALGOLIA_APP_ID,
+  algoliaAPIKey: process.env.ALGOLIA_API_KEY,
+  algoliaIndexName: process.env.ALGOLIA_INDEX_NAME,
+  collectionPath: process.env.COLLECTION_PATH,
+  fields: process.env.FIELDS,
 };
