+++ functions/src/extract.ts
@@ -1,3 +1,4 @@
+'use strict';
 /*
  * Copyright 2021 Algolia
  *
