+++ functions/src/util.ts
@@ -1,0 +1,50 @@
+'use strict';
+/*
+ * Copyright 2021 Algolia
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import * as functions from 'firebase-functions';
+import { DocumentSnapshot } from 'firebase-functions/lib/providers/firestore';
+
+export enum ChangeType {
+  CREATE,
+  DELETE,
+  UPDATE,
+}
+
+export const getChangeType = (
+  change: functions.Change<DocumentSnapshot>
+) => {
+  if (!change.after.exists) {
+    return ChangeType.DELETE;
+  }
+  if (!change.before.exists) {
+    return ChangeType.CREATE;
+  }
+  return ChangeType.UPDATE;
+};
+
+export const buildRequestOptions = () => {
+  return {
+    headers: {
+      'User-Agent': 'firestore_integration_(0.0.1)',
+    },
+  };
+}
+
+export const getObjectSizeInBytes = (object: [] | {}) => {
+  const recordBuffer = Buffer.from(JSON.stringify(object));
+  return recordBuffer.byteLength;
+}
